requirements.txt
Flask>=3.0.0
Flask-CORS>=4.0.0
Flask-Limiter>=3.5.0
Flask-Talisman>=1.1.0
Flask-JWT-Extended>=4.6.0
marshmallow>=3.20.0
bcrypt>=4.1.0
openai>=1.3.0
python-dotenv>=1.0.0
Pillow>=11.0.0
fal-client>=0.4.0
requests>=2.31.0
python-keycloak>=3.9.0
PyJWT>=2.8.0
